Personal Finance Tracker API/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
python-jose
python-dotenv
pydantic[email]
datetime
slowapi
limits
